refactor(frontend): extract recipe modal model, source helpers, summary, footer
526-line modal becomes a 126-line view over useRecipeModalModel (hook
order preserved), pure source/format helpers, and two section views.
